Fixes xgui module not updating groups
This fixes an issue reported by @darkfated that when a new group is created, the UnoLimited would trigger and error because the if statement was not valid, is to "and" and I put a or.
When a new group is create it trigger and repopulation of the groups in the ULX XGUI menu.